+++ lib/config/import_models.d.ts
@@ -2,6 +2,7 @@ export { IClientAuthContext, ClientAuthContext } from '../classes/models/Client/
 export { IClientOAuthAuthenticationContext, ClientOAuthAuthenticationContext } from './../classes/models/Client/ClientOAuthAuthenticationContext';
 export { IClientFacebookAuthenticationContext, ClientFacebookAuthenticationContext } from './../classes/models/Client/ClientFacebookAuthenticationContext';
 export { IClientAppleSignInAuthenticationContext, ClientAppleSignInAuthenticationContext } from './../classes/models/Client/ClientAppleSignInAuthenticationContext';
+export { IClientSimpleAuthenticationData, ClientSimpleAuthenticationData } from './../classes/models/Client/ClientSimpleAuthenticationData';
 export { ClientAccountInformation } from './../classes/models/Client/ClientAccountInformation';
 export { IClientAccountRegisterContext, ClientAccountRegisterContext } from './../classes/models/Client/ClientAccountRegisterContext';
 export { IClientAccountUpdateContext, ClientAccountUpdateContext } from './../classes/models/Client/ClientAccountUpdateContext';